# partB/openapi.yaml
openapi: 3.0.3
info:
  title: URL Shortener API
  version: 0.3.0
  description: Part B checkpoint 3 implementation for a small URL shortener.
servers:
  - url: http://localhost:3000
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /api/links:
    get:
      summary: List saved links
      responses:
        "200":
          description: Link list response
          content:
            application/json:
              schema:
                type: object
                properties:
                  links:
                    type: array
                    items:
                      $ref: "#/components/schemas/Link"
    post:
      summary: Create short URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  format: uri
                  example: https://example.com/very/long/path
                expiresAt:
                  type: string
                  format: date-time
                  nullable: true
                customCode:
                  type: string
                  minLength: 3
                  maxLength: 32
                  pattern: "^[a-zA-Z0-9_-]+$"
                  example: my-link
      responses:
        "201":
          description: Short URL created
          content:
            application/json:
              schema:
                type: object
                properties:
                  link:
                    $ref: "#/components/schemas/Link"
        "400":
          description: Invalid request body
        "409":
          description: Custom short code is already in use
  /api/links/{shortCode}:
    get:
      summary: Get link stats by short code
      parameters:
        - name: shortCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Link stats response
          content:
            application/json:
              schema:
                type: object
                properties:
                  link:
                    $ref: "#/components/schemas/Link"
        "404":
          description: Short URL not found
    delete:
      summary: Delete short link by short code
      parameters:
        - name: shortCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Link deleted
        "404":
          description: Short URL not found
  /{shortCode}:
    get:
      summary: Redirect to original URL
      parameters:
        - name: shortCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirects to the original URL
        "404":
          description: Short URL not found
        "410":
          description: Short URL has expired
components:
  schemas:
    Link:
      type: object
      properties:
        id:
          type: integer
          example: 1
        originalUrl:
          type: string
          example: https://example.com/very/long/path
        shortCode:
          type: string
          example: Ab12Cd3
        shortUrl:
          type: string
          example: http://localhost:3000/Ab12Cd3
        clicks:
          type: integer
          example: 0
        expiresAt:
          type: string
          nullable: true
          example: null
        createdAt:
          type: string
          example: 2026-05-11 12:00:00